+++ Sources/MapboxMaps/Foundation/Extensions/FloatingPoint.swift
@@ -1,0 +1,6 @@
+extension FloatingPoint {
+    internal func wrapped(to range: Range<Self>) -> Self {
+        let d = range.upperBound - range.lowerBound
+        return fmod((fmod((self - range.lowerBound), d) + d), d) + range.lowerBound
+    }
+}